Turns out the sensor did not trigger the alarm due to faulty merged code from my part. It now works. I re-added a delay to the calendar event creation since I would not get the events still. We may have to discuss that because it seems I am missing something:
either the sms event for event creation does not get sent when the event is created by yourself (as opposed to an invitation from somone else)

Moved the Sensor Class in the alarm_model file. This is debatable since the sensor class actually runs its own thread. We may have to discuss this one too.

At least this build runs now.